Fix UBSan "misaligned-pointer-use" warning on aarch64
As `data` type is a pointer to `uint8_t`, the access to the memory via
the `(uint64_t *)data` pointer is misaligned. It can be an issue on ARM
hardware. As suggested in https://en.cppreference.com/w/cpp/language/reinterpret_cast:
"When it is needed to interpret the bytes of an object as a value of a
different type, std::memcpy ... can be used".

For more details, please refer to https://github.com/bitcoin/bitcoin/issues/29178

Also: https://stackoverflow.com/questions/32062894/take-advantage-of-arm-unaligned-memory-access-while-writing-clean-c-code